+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Output the three-dots `...` for the functions `mkostemps()` and
   `mkostempsat()`
 - Remove output the three-dots `...` for the function `umount()`
+- Output the flags in hexadecimal for the functions `mount()`, `nmount()`,
+  `umount2()` and `unmount()`
 
 ### Fixed